requirements.txt
pandas
requests
torch
transformers
datasets
huggingface_hub
python-dotenv
